communication/generic/python/entities/agent/requirements.txt
paho-mqtt>=1.2
MySQL-python>1.2.5
protobuf>=3.1.0
